+++ shoko_api_core/lib/src/_internal.dart
@@ -1,6 +1,5 @@
 export 'data/exceptions/repository_exception.dart';
 
-export 'domain/extensions/exception_extension.dart';
 export 'domain/extensions/request_extension.dart';
 
 export 'domain/utils/error_code.dart';