+++ shoko_api_core/lib/src/data/exceptions/wrong_bearer_token_exception.dart
@@ -1,6 +1,0 @@
-class WrongBearerTokenException implements Exception {
-  @override
-  String toString() {
-    return 'Wrong bearer token Exception';
-  }
-}