+++ shoko_api_core/lib/src/domain/models/route_controller.dart
@@ -1,9 +1,10 @@
 import 'package:dart_frog/dart_frog.dart';
-import 'package:shoko_api_core/src/domain/extensions/exception_extension.dart';
-import 'package:shoko_api_core/src/domain/utils/error_code.dart';
+import 'package:shoko_api_core/src/data/exceptions/repository_exception.dart';
+import 'package:shoko_api_core/src/domain/utils/logger.dart';
+import 'package:shoko_api_core/src/features/backend_responses/backend_error.dart';
 
 abstract class RouteController {
-  Future<Response> direct(Request request, [Map<String, dynamic>? dynamicRouteParams]) async {
+  Future<Response> direct(Request request, {Map<String, dynamic>? dynamicRouteParams}) async {
     try {
       final response = await switch (request.method) {
         HttpMethod.get => get(request, dynamicRouteParams),
@@ -17,17 +18,19 @@ abstract class RouteController {
 
       return response;
     } catch (e, st) {
-      //if e is not an exception then we don't know the error
-      if (e is! Exception) return ErrorCodes.instance.server.undefiend.response;
-      return e.getRresponse(st);
+      if (e is RepositoryException) return e.code.response;
+
+      Logger.instance.e(this, stackTrace: st, stackTraceLength: 4);
+
+      return DefaultCodeResponses.instance.c5XX.c500.response;
     }
   }
 
-  Future<Response> get(Request request, [Map<String, dynamic>? dynamicRouteParams]) async => ErrorCodes.instance.server.notImplemented.response;
-  Future<Response> post(Request request, [Map<String, dynamic>? dynamicRouteParams]) async => ErrorCodes.instance.server.notImplemented.response;
-  Future<Response> put(Request request, [Map<String, dynamic>? dynamicRouteParams]) async => ErrorCodes.instance.server.notImplemented.response;
-  Future<Response> delete(Request request, [Map<String, dynamic>? dynamicRouteParams]) async => ErrorCodes.instance.server.notImplemented.response;
-  Future<Response> head(Request request, [Map<String, dynamic>? dynamicRouteParams]) async => ErrorCodes.instance.server.notImplemented.response;
-  Future<Response> options(Request request, [Map<String, dynamic>? dynamicRouteParams]) async => ErrorCodes.instance.server.notImplemented.response;
-  Future<Response> patch(Request request, [Map<String, dynamic>? dynamicRouteParams]) async => ErrorCodes.instance.server.notImplemented.response;
+  Future<Response> get(Request request, [Map<String, dynamic>? dynamicRouteParams]) async => DefaultCodeResponses.instance.c5XX.c501.response;
+  Future<Response> post(Request request, [Map<String, dynamic>? dynamicRouteParams]) async => DefaultCodeResponses.instance.c5XX.c501.response;
+  Future<Response> put(Request request, [Map<String, dynamic>? dynamicRouteParams]) async => DefaultCodeResponses.instance.c5XX.c501.response;
+  Future<Response> delete(Request request, [Map<String, dynamic>? dynamicRouteParams]) async => DefaultCodeResponses.instance.c5XX.c501.response;
+  Future<Response> head(Request request, [Map<String, dynamic>? dynamicRouteParams]) async => DefaultCodeResponses.instance.c5XX.c501.response;
+  Future<Response> options(Request request, [Map<String, dynamic>? dynamicRouteParams]) async => DefaultCodeResponses.instance.c5XX.c501.response;
+  Future<Response> patch(Request request, [Map<String, dynamic>? dynamicRouteParams]) async => DefaultCodeResponses.instance.c5XX.c501.response;
 }