+++ shoko_api_core/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.0.2
+**Rework error codes to response codes:**
+- Now you can create your own class with ready-made server responses based on `BackendBackendCodeResponses` or user `DefaultCodeResponses`.
+
+**Remove:**
+- WrongBearerTokenException
+- ExceptionExtension
+
 ## 1.0.1
 - Update docs
 - Add `request.validatedBody` and `request.validatedQueryParameters`